tests/wpt/web-platform-tests/app-history/navigate/return-value/forward-beforeunload.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>

<iframe id="i" src="/common/blank.html"></iframe>

<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  assert_equals(i.contentWindow.appHistory.entries().length, 1);
  let key = i.contentWindow.appHistory.current.key;

  i.contentWindow.appHistory.navigate("?1");
  await new Promise(resolve => i.onload = resolve);

  i.contentWindow.appHistory.back();
  await new Promise(resolve => i.onload = resolve);

  assert_equals(i.contentWindow.appHistory.entries().length, 2);
  assert_equals(i.contentWindow.appHistory.current, i.contentWindow.appHistory.entries()[0]);

  let navigateEventCount = 0;
  i.contentWindow.appHistory.onnavigate = () => navigateEventCount++;
  i.contentWindow.appHistory.onnavigatesuccess = t.unreached_func("onnavigatesuccess should not be called");
  i.contentWindow.appHistory.onnavigateerror = t.unreached_func("onnavigateerror should not be called");

  let assertionPromise;
  // The iframe does not have sticky activation, so per
  // https://html.spec.whatwg.org/#prompt-to-unload-a-document, no prompt is
  // shown and the navigation will proceed.
  i.contentWindow.onbeforeunload = t.step_func(() => {
    assertionPromise = assertBothRejectDOM(t, i.contentWindow.appHistory.forward(), "InvalidStateError", i.contentWindow);
  });
  i.contentWindow.appHistory.navigate("?1");

  assert_not_equals(assertionPromise, undefined);
  await assertionPromise;

  assert_equals(navigateEventCount, 1);
}, "forward() inside onbeforeunload");
</script>
